css/cssom-view/matchMediaAddListener-handleEvent.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Test: CSSOM View matchMedia handleEvent via addListener</title>
<link rel="help" href="https://dom.spec.whatwg.org/#callbackdef-eventlistener">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/matchMedia.js"></script>
<div id="log"></div>
<script>
setup({ allow_uncaught_exception: true });

promise_test(async t => {
    const mql = await createMQL(t);

    let _this;
    let _event;
    const eventListener = {
        handleEvent(event) {
            _this = this;
            _event = event;
        },
    };

    mql.addListener(eventListener);
    triggerMQLEvent(mql);
    await waitForChangesReported();

    assert_equals(_this, eventListener);
    assert_equals(_event.media, mql.media);
    assert_equals(_event.matches, mql.matches);
}, "calls handleEvent method of event listener");

promise_test(async t => {
    const mql = await createMQL(t);

    let calls = 0;
    mql.addListener({
        get handleEvent() {
            calls++;
            return function() {};
        },
    });
    assert_equals(calls, 0);

    triggerMQLEvent(mql);
    await waitForChangesReported();
    assert_equals(calls, 1);

    triggerMQLEvent(mql);
    await waitForChangesReported();
    assert_equals(calls, 2);
}, "looks up handleEvent method on every event dispatch");

promise_test(async t => {
    const mql = await createMQL(t);

    let calls = 0;
    const eventListener = function() { calls++; };
    Object.defineProperty(eventListener, "handleEvent", {
        get: t.unreached_func("handleEvent method should not be looked up on functions"),
    });
    mql.addListener(eventListener);

    triggerMQLEvent(mql);
    await waitForChangesReported();
    assert_equals(calls, 1);
}, "doesn't look up handleEvent method on callable event listeners");

const uncaught_error_test = async (t, listener) => {
    const mql = await createMQL(t);
    mql.addListener(listener);

    const eventWatcher = new EventWatcher(t, window, "error");
    const errorPromise = eventWatcher.wait_for("error");
    triggerMQLEvent(mql);

    const event = await errorPromise;
    throw event.error;
};

promise_test(t => {
    const error = { name: "test" };
    const listener = {
        get handleEvent() {
            throw error;
        },
    };

    return promise_rejects_exactly(t, error, uncaught_error_test(t, listener));
}, "rethrows errors when getting handleEvent");

promise_test(t => {
    const listener = { handleEvent: null };
    return promise_rejects(t, new TypeError(), uncaught_error_test(t, listener));
}, "throws if handleEvent is falsy and not callable");

promise_test(t => {
    const listener = { handleEvent: "str" };
    return promise_rejects(t, new TypeError(), uncaught_error_test(t, listener));
}, "throws if handleEvent is thruthy and not callable");
</script>
